+++ python/taskito/contrib/__init__.py
@@ -1,0 +1,1 @@
+"""taskito integrations with third-party frameworks."""